onepad-legacy: Use mt_queue, and move it from onepad to common.
