+++ plugins/onepad_legacy/onepad.cpp
@@ -53,11 +53,8 @@ const u32 build = 3; // increase that with each version
 
 FILE *padLog = NULL;
 
-pthread_spinlock_t mutex_KeyEvent;
-bool mutex_WasInit = false;
 KeyStatus *key_status = NULL;
-
-std::queue<keyEvent> ev_fifo;
+MtQueue<keyEvent> g_ev_fifo;
 
 static void InitLibraryName()
 {
@@ -193,11 +190,7 @@ EXPORT_C_(s32) PADopen(void *pDsp)
 {
     memset(&event, 0, sizeof(event));
     key_status->Init();
-
-    while (!ev_fifo.empty())
-        ev_fifo.pop();
-    pthread_spin_init(&mutex_KeyEvent, PTHREAD_PROCESS_PRIVATE);
-    mutex_WasInit = true;
+    g_ev_fifo.reset();
 
 #if defined(__unix__)
     GamePad::EnumerateGamePads(s_vgamePad);
@@ -223,10 +216,6 @@ EXPORT_C_(void) PADsetLogDir(const char *dir)
 
 EXPORT_C_(void) PADclose()
 {
-    while (!ev_fifo.empty())
-        ev_fifo.pop();
-    mutex_WasInit = false;
-    pthread_spin_destroy(&mutex_KeyEvent);
     _PADclose();
 }
 
@@ -349,13 +338,6 @@ EXPORT_C_(keyEvent *) PADkeyEvent()
 #if defined(__unix__)
 EXPORT_C_(void) PADWriteEvent(keyEvent &evt)
 {
-    // This function call be called before PADopen. Therefore we cann't
-    // guarantee that the spin lock was initialized
-    if (mutex_WasInit)
-    {
-        pthread_spin_lock(&mutex_KeyEvent);
-        ev_fifo.push(evt);
-        pthread_spin_unlock(&mutex_KeyEvent);
-    }
+    g_ev_fifo.push(evt);
 }
 #endif
